requirements.txt
telethon>=1.42.0
python-dotenv==1.1.1
motor>=3.3.0
dnspython>=2.6.0
